requirements.txt
PILLOW
flask_sqlalchemy
flask
datetime
sqlalchemy
flask_migrate